Fix the path issue in the tests
I'm not proud of this one, but it's hard to modify the entire test suite to accomodate output path of the commands during testing. This way I am cheating a bit, and pointing to the real file that will be copied over, so no harm done. Removing the 'wp-content' folder should be an ask probably.