Update and add tests
Add test for the core download verification, add fix for plugin install failure (force parameter was needed to verify that the installation will pass). Probably need to add the same but without to verify a failure will happen. Some tests are still failig, need to see why.

The test for dropin folder removal is skipped because we cannot run Pest tests in process isolation (in v1 which supports PHP 7.4).