Update the init command
This update added the --force parameter to override the WordPress core install.
Also, I removed some directory copy methods in favor of the filesystem component which works better.
I also needed to fix the failure that happened because new WP version has additional files that
wordpress-develop doesn't have, so now we need to download both the core and the develop folders.

Still need to test if this will solve the problem I was experiencing on the new WP versions.